split off ClassMetadataInfo from ClassMetadata (still missing some methods to comply to the ODM interface)
